Great threads handling and performance improvements:
 - io_context not needed anymore

 - Now threads pool can be per session

 - Each device now can have own type of threads pool with own threads count

+ docs updates and some cleanups and minor fixes.



git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@1563 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ iscsi-scst/usr/iscsi_scstd.c
@@ -583,11 +583,11 @@ static void init_max_params(void)
 	}
 
 	/* QueuedCommands */
-	target_keys[key_queued_cmnds].local_def = min(target_keys[key_queued_cmnds].local_def,
+	target_keys[key_queued_cmnds].local_def = min((int)target_keys[key_queued_cmnds].local_def,
 						      iscsi_init_params.max_queued_cmds);
-	target_keys[key_queued_cmnds].max = min(target_keys[key_queued_cmnds].max,
+	target_keys[key_queued_cmnds].max = min((int)target_keys[key_queued_cmnds].max,
 						iscsi_init_params.max_queued_cmds);
-	target_keys[key_queued_cmnds].min = min(target_keys[key_queued_cmnds].min,
+	target_keys[key_queued_cmnds].min = min((int)target_keys[key_queued_cmnds].min,
 						iscsi_init_params.max_queued_cmds);
 
 	/* MaxRecvDataSegmentLength */
